requirements.txt
psycopg2-binary
uvicorn[standard]
sqlalchemy
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fastapi>=0.95.0
aiofiles
sqlalchemy-utils
pytz
requests
aiohttp
